requirements.txt
Flask==3.0.3
gunicorn==23.0.0
flickrapi==2.4.0
python-dotenv==1.0.1
requests==2.32.3
